requirements.txt
streamlit
wordcloud
matplotlib
qrcode
pillow
PyGithub
plotly